+++ client/src/lib.rs
@@ -4,6 +4,7 @@ pub mod backend;
 pub mod backend_fs;
 pub mod backend_http;
 pub mod config;
+pub mod container;
 pub mod keychain;
 pub mod token;
 pub mod types;